remove logs, move out non-pub fn-s
